Conflicts:
	Code/projet_lo21-build-desktop-Qt_4_8_0_for_Desktop_-_MinGW__Qt_SDK__Debug/debug/mainwindow.o
	Code/projet_lo21-build-desktop-Qt_4_8_0_for_Desktop_-_MinGW__Qt_SDK__Debug/debug/projet_lo21.exe
	Code/projet_lo21-build-desktop-Qt_4_8_0_for_Desktop_-_MinGW__Qt_SDK__Debug/ui_mainwindow.h
	Code/projet_lo21/projet_lo21.pro.user

+++ Code/projet_lo21/mainwindow.cpp
@@ -297,9 +297,9 @@ void MainWindow::createOnglet()
 
 }
 
-void MainWindow::on_tabWidget_currentChanged(unsigned int index)
+void MainWindow::on_tabWidget_currentChanged(int index)
 {
-    if(index>=Collection_pile::getInstance().size())
+    if((unsigned int)index>=Collection_pile::getInstance().size())
     createOnglet();
     Collection_pile::getInstance().setActif(index);
     emit refresh_signal();
@@ -448,3 +448,4 @@ void MainWindow::on_nonClavier_clicked(){
     resize(size);
 
 }
+
